backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.19.0
PyJWT==2.8.0
email-validator==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
sib-api-v3-sdk==7.6.0 